make-dist: Fetch package index file
Let the make-dist script bootstrap rebar3 to fetch the package index
file.  Running the bootstrap script from the generated archive without a
package index would fail.

Now that rebar3 is built anyway, it could also be used to fetch the
dependencies.  However, we'll stick to the existing code, at least for
the moment. This gives us full control over checking the list of
downloaded apps, and their licenses.